added some error handling in many js files

+++ server/schemas/resolvers.js
@@ -37,19 +37,25 @@ const resolvers = {
     },
     //find owner of tool
     toolOwner: async (parents, args, context) => {
-      //find user if it contains the checkout ID
-      // console.log('args', args);
-      const owner = await User.findOne({ myTools: args._id }).populate(
-        "borrowedTools myTools"
-      );
-      // console.log("resolver line 37", owner);
-      return owner;
+      if (context.user) {
+        //find user if it contains the checkout ID
+        // console.log('args', args);
+        const owner = await User.findOne({ myTools: args._id }).populate(
+          "borrowedTools myTools"
+        );
+        // console.log("resolver line 37", owner);
+        return owner;
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
     //views specific tools
     tool: async (parent, args, context) => {
-      const tool = await Tool.findById(args).populate("checkout");
+      if (context.user) {
+        const tool = await Tool.findById(args).populate("checkout");
 
-      return tool;
+        return tool;
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
     myTools: async (parent, args, context) => {
       if (context.user) {
@@ -59,10 +65,13 @@ const resolvers = {
     },
     // A query that returns all of the tool objects assosciate with an array of checkout ids, used in succesion with the myTools query
     borrowedTools: async (parent, args, context) => {
-      const tool = await Tool.find(
-        { checkout: { $in: args.idArray } }).populate('checkout');
-      console.log(tool);
-      return tool;
+      if (context.user) {
+        const tool = await Tool.find(
+          { checkout: { $in: args.idArray } }).populate('checkout');
+        console.log(tool);
+        return tool;
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
     //views all the tools
     tools: async () => {
@@ -74,51 +83,59 @@ const resolvers = {
     },
     garage: async (parent, args, context) => {
       // const garage = await Garage.findById(args).populate('admin').populate('members.myTools.name').populate('members.myTools.checkout').populate('admin');
-
-      const garage = await Garage.findById(args)
-        .populate("admin")
-        .populate({
-          path: "members",
-          model: "User",
-          populate: {
-            path: "myTools",
-            model: "Tool",
+      if (context.user) {
+        const garage = await Garage.findById(args)
+          .populate("admin")
+          .populate({
+            path: "members",
+            model: "User",
             populate: {
-              path: "checkout",
-              model: "Checkout",
+              path: "myTools",
+              model: "Tool",
+              populate: {
+                path: "checkout",
+                model: "Checkout",
+              },
             },
-          },
-        });
+          });
 
-      return garage;
+        return garage;
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
     // find the user connected to a checkout
     checkoutBorrower: async (parents, args, context) => {
-      const borrower = await User.findOne({ borrowedTools: args._id });
-      return borrower;
+      if (context.user) {
+        const borrower = await User.findOne({ borrowedTools: args._id });
+        return borrower;
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
     checkout: async (parents, args, context) => {
-      return await Checkout.findOne({ _id: args._id });
+      if (context.user) {
+        return await Checkout.findOne({ _id: args._id });
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
 
     messages: async (parent, args, context) => {
-      // if (context.user) {
-      // console.log("MESSAGES ARGS ", args);
-      const garageWithMessages = await Garage.findById(args._id)
-        .populate({
-          path: "messages",
-          model: "Message",
-          populate: {
-            path: "author",
-            model: "User",
-          }
-        });
-      // console.log("MESSAGES GARAGE ", garageWithMessages);
-      return garageWithMessages;
-      // }
-      // throw new AuthenticationError("You need to be logged in!");
+      if (context.user) {
+        // console.log("MESSAGES ARGS ", args);
+        const garageWithMessages = await Garage.findById(args._id)
+          .populate({
+            path: "messages",
+            model: "Message",
+            populate: {
+              path: "author",
+              model: "User",
+            }
+          });
+        // console.log("MESSAGES GARAGE ", garageWithMessages);
+        return garageWithMessages;
+      }
+      throw new AuthenticationError("You need to be logged in!");
     },
-  
+
   },
   Mutation: {
     // creates user
@@ -149,57 +166,57 @@ const resolvers = {
     // updates user's name, phone and address, email and password cannot be changed
     // front end must always set the value to what was originally there so a field wont be nulled accidentally
     updateUser: async (parent, args) => {
-      // if (context.user) {
-      console.log("PARENT");
-      console.log(parent);
-      console.log("ARGS");
-      console.log(args);
-      return await User.findByIdAndUpdate(
-        { _id: args._id },
-        {
-          name: args.name,
-          phone: args.phone,
-          address: args.address,
-          email: args.email,
-        },
-        { new: true, runValidators: true }
-      );
-      // }
+      if (context.user) {
+        console.log("PARENT");
+        console.log(parent);
+        console.log("ARGS");
+        console.log(args);
+        return await User.findByIdAndUpdate(
+          { _id: args._id },
+          {
+            name: args.name,
+            phone: args.phone,
+            address: args.address,
+            email: args.email,
+          },
+          { new: true, runValidators: true }
+        );
+      }
       // If user attempts to execute this mutation and isn't logged in, throw an error
-      // throw new AuthenticationError('You need to be logged in!');
+      throw new AuthenticationError('You need to be logged in!');
     },
     // removes user
     removeUser: async (parent, args, context) => {
-      // if (context.user) {
-      const user = await User.findByIdAndDelete(args).populate("myTools");
-      //also delete the associated tools
-      await Tool.deleteMany({ _id: { $in: user.myTools } });
+      if (context.user) {
+        const user = await User.findByIdAndDelete(args).populate("myTools");
+        //also delete the associated tools
+        await Tool.deleteMany({ _id: { $in: user.myTools } });
 
-      //find list of all garages that this user is admin of
-      const adminGarages = await Garage.find({ admin: args._id });
+        //find list of all garages that this user is admin of
+        const adminGarages = await Garage.find({ admin: args._id });
 
-      //delete all the garages this user is admin of
-      const deletedGarage = await Garage.deleteMany({ admin: args._id });
+        //delete all the garages this user is admin of
+        const deletedGarage = await Garage.deleteMany({ admin: args._id });
 
-      //for the removed garages remove it from users who have it as one of their garages
-      adminGarages.forEach(async (garage) => {
-        //for each garage deleted remove it from the user's array
-        await User.updateMany(
-          { myGarages: garage._id },
-          { $pull: { myGarages: garage._id } }
+        //for the removed garages remove it from users who have it as one of their garages
+        adminGarages.forEach(async (garage) => {
+          //for each garage deleted remove it from the user's array
+          await User.updateMany(
+            { myGarages: garage._id },
+            { $pull: { myGarages: garage._id } }
+          );
+        });
+
+        // remove the user from garages that the user is a part off
+        await Garage.updateMany(
+          { members: args._id },
+          { $pull: { members: args._id } },
+          { new: true }
         );
-      });
 
-      // remove the user from garages that the user is a part off
-      await Garage.updateMany(
-        { members: args._id },
-        { $pull: { members: args._id } },
-        { new: true }
-      );
-
-      return user;
-      // }
-      // throw new AuthenticationError('You need to be logged in!');
+        return user;
+      }
+      throw new AuthenticationError('You need to be logged in!');
     },
 
     // user can add tool
@@ -227,21 +244,22 @@ const resolvers = {
     },
     // user removes tool
     removeTool: async (parent, args, context) => {
-      // if (context.user) {
-      // delete tool but capture its data
-      const removedTool = await Tool.findByIdAndDelete(args);
+      if (context.user) {
+        // delete tool but capture its data
+        const removedTool = await Tool.findByIdAndDelete(args);
 
-      //remove the tool from the user (one to one relationship)
-      return await User.findOneAndUpdate(
-        { myTools: args._id },
-        { $pull: { myTools: args._id } },
-        { new: true }
-      ).populate("myTools");
-      // }
-      // throw new AuthenticationError('You need to be logged in!');
+        //remove the tool from the user (one to one relationship)
+        return await User.findOneAndUpdate(
+          { myTools: args._id },
+          { $pull: { myTools: args._id } },
+          { new: true }
+        ).populate("myTools");
+      }
+      throw new AuthenticationError('You need to be logged in!');
     },
 
     updateTool: async (parent, args, context) => {
+      if (context.user) {
       const updatedTool = await Tool.findOneAndUpdate(
         { _id: args._id },
         {
@@ -254,6 +272,8 @@ const resolvers = {
       )
 
       return updatedTool;
+    }
+    throw new AuthenticationError('You need to be logged in!');
     },
     // garage creation accepts userId (atm ID but TODO: login details), garageName and description
     // create Garage and return it with admin and members populated, the user who created it automatically becomes a member
@@ -267,7 +287,7 @@ const resolvers = {
           members: [context.user._id],
         });
 
-        const garage = await Garage.findById(newGarage._id )       
+        const garage = await Garage.findById(newGarage._id)
           .populate("admin")
           .populate("members");
 
@@ -319,12 +339,12 @@ const resolvers = {
 
         // console.log('GARAGE USER resolver', users);
 
-        const chatIDs=garage.messages;
+        const chatIDs = garage.messages;
 
-        console.log("DELETE GARAGE resolver",chatIDs);
+        console.log("DELETE GARAGE resolver", chatIDs);
 
         // TODO: after deleting garage, parse through it to delete messages. This is done when messages are done.
-        const messages= await Message.deleteMany({_id:{$in:chatIDs}});
+        const messages = await Message.deleteMany({ _id: { $in: chatIDs } });
 
         console.log("messages in delete garage", messages)
 
@@ -333,8 +353,6 @@ const resolvers = {
       }
       throw new AuthenticationError('You need to be logged in and have admin status to delete this garage!');
     },
-
-    //TODO: Do UPDATE the Garage. Only the admin can update the garage
 
     // join garage, find the garage by invitation code and update its members array with member
     //update member's myGarages with the garage ID
@@ -380,65 +398,69 @@ const resolvers = {
     },
 
     addCheckout: async (parent, args, context) => {
-      // if (context.user) {
-      const checkout = await Checkout.create({
-        outDate: args.outDate,
-        dueDate: args.dueDate,
-        approved: false
-      });
+      if (context.user) {
+        const checkout = await Checkout.create({
+          outDate: args.outDate,
+          dueDate: args.dueDate,
+          approved: false
+        });
 
-      await User.findOneAndUpdate(
-        { _id: context.user._id },
-        { $addToSet: { borrowedTools: checkout._id } },
-        { new: true }
-      ).populate("borrowedTools");
+        await User.findOneAndUpdate(
+          { _id: context.user._id },
+          { $addToSet: { borrowedTools: checkout._id } },
+          { new: true }
+        ).populate("borrowedTools");
 
-      return await Tool.findOneAndUpdate(
-        { _id: args.toolId },
-        { checkout: checkout },
-        { new: true }
-      ).populate("checkout");
-      // }
-      // throw new AuthenticationError('You need to be logged in!');
+        return await Tool.findOneAndUpdate(
+          { _id: args.toolId },
+          { checkout: checkout },
+          { new: true }
+        ).populate("checkout");
+      }
+      throw new AuthenticationError('You need to be logged in!');
     },
 
     approveCheckout: async (parent, args, context) => {
+      if (context.user) {
       return await Checkout.findOneAndUpdate(
         { _id: args._id },
         { $set: { approved: true } }
       );
+    }
+    throw new AuthenticationError('You need to be logged in!');
     },
 
     // user removes tool
     deleteCheckout: async (parent, args, context) => {
-      // if (context.user) {
-      const parentTool = await Tool.findOne({ _id: args._id });
-      await Checkout.findByIdAndDelete(parentTool.checkout._id);
+      if (context.user) {
+        const parentTool = await Tool.findOne({ _id: args._id });
+        await Checkout.findByIdAndDelete(parentTool.checkout._id);
 
-      await User.findOneAndUpdate(
-        { borrowedTools: context.user._id },
-        { $pull: { borrowedTools: args._id } },
-        { new: true },
-      ).populate('borrowedTools');
+        await User.findOneAndUpdate(
+          { borrowedTools: context.user._id },
+          { $pull: { borrowedTools: args._id } },
+          { new: true },
+        ).populate('borrowedTools');
 
 
-      return await Tool.findOneAndUpdate(
-        { _id: args._id },
-        { checkout: null },
-        { new: true }
-      ).populate('checkout');
-      // throw new AuthenticationError('You need to be logged in!');
+        return await Tool.findOneAndUpdate(
+          { _id: args._id },
+          { checkout: null },
+          { new: true }
+        ).populate('checkout');
+      }
+      throw new AuthenticationError('You need to be logged in!');
     },
 
     addMessage: async (parent, args, context) => {
       if (context.user) {
         // create the message
-       const newMessage = await Message.create({
+        const newMessage = await Message.create({
           author: context.user._id,
           body: args.body,
         });
 
-        const message= await Message.findById(newMessage._id).populate('author');
+        const message = await Message.findById(newMessage._id).populate('author');
 
         // push message to garage
         const garage = await Garage.findByIdAndUpdate(
@@ -453,8 +475,11 @@ const resolvers = {
 
     //delete Message
     deleteMessage: async (parent, args, context) => {
-      const delMessage = await Message.findByIdAndDelete(args._id);
-      return delMessage;
+      if (context.user) {
+        const delMessage = await Message.findByIdAndDelete(args._id);
+        return delMessage;
+      }
+      throw new AuthenticationError('You need to be logged in!');
     },
 
   },
